Small fixes

+++ app/src/main/java/com/mypedometer/pip/pedometer/DataBaseHelper.java
@@ -421,14 +421,14 @@ public class DataBaseHelper extends SQLiteOpenHelper {
                 user.setIsPrivateProfile(Boolean.parseBoolean(cursorInfo.getString(7)));
 
                 user.setTotalSteps(Integer.parseInt(cursorStats.getString(1)));
-                user.setTotalDistance(Integer.parseInt(cursorStats.getString(2)));
-                user.setTotalCalories(Integer.parseInt(cursorStats.getString(3)));
+                user.setTotalDistance(Float.parseFloat(cursorStats.getString(2)));
+                user.setTotalCalories(Float.parseFloat(cursorStats.getString(3)));
                 user.setDailySteps(cursorStats.getString(4));
                 user.setDailyDistance(cursorStats.getString(5));
                 user.setDailyCalories(cursorStats.getString(6));
                 user.setGoalDailySteps(Integer.valueOf(cursorStats.getString(7)));
-                user.setGoalDailyDistance(Integer.valueOf(cursorStats.getString(8)));
-                user.setGoalDailyCalories(Integer.valueOf(cursorStats.getString(9)));
+                user.setGoalDailyDistance(Float.valueOf(cursorStats.getString(8)));
+                user.setGoalDailyCalories(Float.valueOf(cursorStats.getString(9)));
                 user.setChallengeSteps(cursorStats.getString(10));
                 user.setChallengeDistance(cursorStats.getString(11));
                 user.setChallengeCalories(cursorStats.getString(12));
